requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-widget-tweaks==1.5.0
Pillow==10.0.1
python-decouple==3.8
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0